W122 salvage: strip stale markers on _h_camp_data FUN_00167f40 (248/256 nd0, also clears over-window) and frFont FUN_003B1A90 (108/112 nd0) to bank both MATCHes

+++ src/comuTimerSequence.c
@@ -110,7 +110,7 @@ u32 FUN_003bf720(void)
   int *piVar3;
   int lVar5;
   u32 uVar4;
-  u32 bVar2;
+  u8 bVar2;
   u8 auStack_30[0x1c];
 
   piVar3 = (int *)FUN_00195540();
